refactor(cockpit): extract _RailListView into sibling module (#1354)
Eighth wedge of the cockpit_ui.py god-module split. _RailListView is the
small rail-aware ListView subclass (~60 LOC) that re-resolves orphaned
click events against the live _nodes list (#964 regression guard). Only
depends on Textual primitives (ListView, ListItem), has a single call
site inside cockpit_ui, and one test that imports via the public path —
clean self-contained wedge.

Moved into new sibling src/pollypm/cockpit_rail_listview.py. Left a
re-export shim in cockpit_ui so existing import paths
(pollypm.cockpit_ui._RailListView) keep working unchanged.

cockpit_ui.py: 18,678 -> 18,620 LOC (-58)
Diff inside cockpit_ui.py: 3 inserts / 61 deletes (well under 200 budget)
New module: 87 LOC including docstrings.

Tests: 196 cockpit tests pass; the #964 regression test
(test_rail_listview_swallows_value_error_for_orphan_clicks) still
imports via the public path and passes.
